Wired up existing export buttons to new /backup endpoint. Renamed and refactored backend code
