--- service/src/main/java/uk/gov/mca/beacons/api/export/csv/SpreadsheetDTO.java
@@ -1,9 +1,0 @@
-package uk.gov.mca.beacons.api.export.csv;
-
-import java.util.List;
-import uk.gov.mca.beacons.api.export.SpreadsheetRow;
-
-public class SpreadsheetDTO {
-
-  public List<SpreadsheetRow> spreadsheetRows;
-}
