+++ service/src/test/java/uk/gov/mca/beacons/api/export/JsonSerialiserUnitTest.java
@@ -2,15 +2,14 @@ package uk.gov.mca.beacons.api.export;
 
 import static org.junit.Assert.assertEquals;
 
-import io.swagger.v3.core.util.Json;
 import java.time.LocalDateTime;
 import java.time.Month;
 import java.util.List;
 import net.minidev.json.JSONArray;
 import net.minidev.json.JSONObject;
 import org.junit.jupiter.api.Test;
-import uk.gov.mca.beacons.api.export.csv.JsonSerialiser;
 import uk.gov.mca.beacons.api.export.rest.*;
+import uk.gov.mca.beacons.api.export.xlsx.backup.JsonSerialiser;
 import uk.gov.mca.beacons.api.shared.rest.person.dto.AddressDTO;
 
 public class JsonSerialiserUnitTest {
